Update `requirements.txt` to use `browser-use==0.7.10` and refactor imports across multiple files for improved compatibility. Removed deprecated imports and adjusted code to align with the new structure of `browser_use`.
