+++ requirements.txt
@@ -1,4 +1,4 @@
-browser-use==0.6.1
+browser-use==0.7.10
 pyperclip==1.9.0
 gradio==5.43.1
 json-repair==0.49.0
